alimentos/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
mysql-connector-python==9.1.0
pydantic==2.10.3
python-dotenv==1.0.1
pydantic-settings==2.6.1